Added laravel sanctum

+++ routes/api.php
@@ -9,20 +9,20 @@ use App\Http\Controllers\ReviewController;
 
 Route::get('/availabilities/{id}', [AvailabilityController::class, 'index']);
 
-Route::post('/availabilities/{id}', [AvailabilityController::class, 'storeEvents']);
+Route::post('/availabilities/{id}', [AvailabilityController::class, 'storeEvents'])->middleware(['auth:sanctum', 'ownOrAdmin']);
 
 Route::get('/exceptions/{id}', [ExceptionController::class, 'index']);
 
-Route::post('/exceptions/{id}', [ExceptionController::class, 'storeExceptions']);
+Route::post('/exceptions/{id}', [ExceptionController::class, 'storeExceptions'])->middleware(['auth:sanctum', 'ownOrAdmin']);
 
-Route::put('/exceptions/{id}', [ExceptionController::class, 'update']);
+Route::put('/exceptions/{id}', [ExceptionController::class, 'update'])->middleware(['auth:sanctum', 'ownOrAdmin']);
 
-Route::delete('/exceptions/{id}', [ExceptionController::class, 'destroy']);
+Route::delete('/exceptions/{id}', [ExceptionController::class, 'destroy'])->middleware(['auth:sanctum', 'ownOrAdmin']);
 
 Route::get('/map/teachers', [MapController::class, 'index']);
 
 // Reviews routes
-Route::post('/reviews', [ReviewController::class, 'saveReview']);
+Route::post('/reviews', [ReviewController::class, 'saveReview'])->middleware(['auth:sanctum', 'role:user']);
 Route::get('/teachers/{teacher}/reviews', [ReviewController::class, 'getTeacherReviews']);
 
 // Chat routes
